simulacao com velocidade cosntante

+++ src/main/java/br/unb/cic/bd/simuladortrafego/arco/Arco.java
@@ -2,15 +2,15 @@ package br.unb.cic.bd.simuladortrafego.arco;
 
 public class Arco {
 	
-	public static final float VELOCIDADE_MEDIA_DEFAULT = 60;
+	public static final double VELOCIDADE_MEDIA_DEFAULT = 60;
 	
 	private String nome;
 	private String linha;
-	private float velocidadeMedia;
-	private float tamanho;
+	private double velocidadeMedia;
+	private double tamanho;
 	private int numero;
 
-	public Arco(String linha, String nome, float velocidadeMedia, float tamanho) {
+	public Arco(String linha, String nome, double velocidadeMedia, double tamanho) {
 		super();
 		this.nome = nome;
 		this.linha = linha;
@@ -19,7 +19,7 @@ public class Arco {
 		this.numero = Integer.parseInt(nome.substring(1));
 	}
 
-	public Arco(String linha, String nome, float tamanho) {
+	public Arco(String linha, String nome, double tamanho) {
 		super();
 		this.linha = linha;
 		this.nome = nome;
@@ -44,19 +44,19 @@ public class Arco {
 		this.nome = nome;
 	}
 
-	public float getVelocidadeMedia() {
+	public double getVelocidadeMedia() {
 		return velocidadeMedia;
 	}
 
-	public void setVelocidadeMedia(float velocidadeMedia) {
+	public void setVelocidadeMedia(double velocidadeMedia) {
 		this.velocidadeMedia = velocidadeMedia;
 	}
 
-	public float getTamanho() {
+	public double getTamanho() {
 		return tamanho;
 	}
 
-	public void setTamanho(float tamanho) {
+	public void setTamanho(double tamanho) {
 		this.tamanho = tamanho;
 	}
